Many changes:
(gtags-mode--roots-list) : Renamed to gtags-mode--alist. It is now a list
of plists.
(gtags-mode--has-open-root) : Renamed to gtags-mode--get-plist. It now
returns the plist associated with root after appliying the action.
(gtags-mode--exec-async-sentinel) : Now executes an action to clear the
completions cache in the associated plist.
(gtags-mode--exec-async) : Use process-put to remember the caller
buffer.
(gtags-mode--list-completions) : New function to return the completion
list associated with a string, or return-generate the cache
conveniently.
(gtags-mode--find-file-hook) : Update to new functions
(xref-backend-identifier-completion-table) : Use the new function
gtags-mode--list-completions.
(project-files) : Delete duplicates when multiple outputs.
(gtags-mode-completion-function) : New function to provide a capf
backend.