Use gtags-mode--local-plist in project-find-functions
(gtags-mode--local-plist) : Receives dir as new optional argument to set
default argument before setting the gtags-mode--plist.  The argument is
required by project api, but in most cases it will be default-directory;
but it needs to be optional because the same function is used by
xref-backend-functions that doesn't use pass any argument.