Unquote arguments and simplify parameters
The quoting target argument for chars like '=' are wrong for gtags
because when passing '\=' global recognizes the \ as a regex char.  This
makes the gtags-mode--quote function more a problem/issue than a
solution. SO this commit attempts to remove it.

When removing that function the whole function call interfaces can be
significantly simplified. So, thats what I am doing here as well.  The
new functions receive the arguments as a &rest array; and there is not
division between arguments and target.