Fix small performance issue
Use gtags-mode--local-plist in project-find-functions
This avoids the excessive call to file-truename (which seems a bit
expensive specially in MS-windows)
(gtags-mode--set-local-plist) : New function with the code in the else
condition of gtags-mode--local-plist. This is the one calling
file-truename now.
(gtags-mode--local-plist) : Use the new function
(gtags-mode-project-backend) : Removed function use
gtags-mode--local-plist instead.
(gtags-mode--after-save-hook) : New condition that uses
gtags-mode--plist, but still attempts to set it.