Segregate the root search function
Fixes #7
This is needed to suppress the annoying warning messages when looking
for a root.
Also to simplify and group the logic in the gtags-mode--create-plist
function.
At the same time this follows (partially) the single responsibility
principle.
(gtags-mode--get-root): New inlined function
(gtags-mode--create-plist): Use the new function and remove the inline
logics.