Correct path to file when definition is found in GTAGSLIBPATH
--path-style=through always prints the path as it is found in the GPATH file,
which is not necessarily equivalent to "relative from the project root".

See https://lists.gnu.org/archive/html/bug-global/2025-05/msg00000.html and
https://lists.gnu.org/archive/html/bug-global/2025-05/msg00001.html

If you want to always print the path as relative to the root of the project, it
is better to use --path-style=relative (which can be omitted, being default) in
conjunction with -C or --directory pointing to the project root.